// IguideME.Web/wwwroot-old/src/components/StudentDashboard/types.ts
import { CanvasStudent } from "../../models/canvas/Student";
import {Tile, AssignmentSubmission} from "../../models/app/Tile";
import {CanvasDiscussion} from "../../models/canvas/Discussion";
import {LearningOutcome} from "../../models/app/LearningGoal";
import { PredictedGrade } from "../../models/app/PredictiveModel";
import { HistoricTileGrades } from "./TileHistoricGraph/types";

export interface IProps {
  student: CanvasStudent | undefined;
  consent?: string | null
}

export type ViewTypes = "bar" | "grid";

export interface IState {
  tilesGradeSummary: TilesGradeSummary[];
  peerGrades: PeerGrades[];
  loaded: boolean;
  displayTile: Tile | null;
  viewType: ViewTypes;
  learningOutcomes: LearningOutcome[];
  discussions: CanvasDiscussion[];
  userSubmissions: Map<number, AssignmentSubmission[]>;
  predictions: PredictedGrade[];
  goalGrade: number;
  settings_view: boolean;
  historicGrades: Map <number,HistoricTileGrades>;
}

export interface TilesGradeSummary {
  tile: Tile, average: number
}

export interface PeerGrades {
  min: number, max: number, avg: number, tileID: number
}